// packages/api/callbreak/src/callbreak.constants.ts
import { type Config, names } from "unique-names-generator";

export enum GameEvents {
	PLAYER_JOINED = "player-joined",
	ALL_PLAYERS_JOINED = "all-players-joined",
	DEAL_CREATED = "deal-created",
	DEAL_WIN_DECLARED = "deal-win-declared",
	ALL_DEAL_WINS_DECLARED = "all-deal-wins-declared",
	ROUND_CREATED = "round-created",
	CARD_PLAYED = "card-played",
	ROUND_COMPLETED = "round-completed",
	DEAL_COMPLETED = "deal-completed",
	STATUS_UPDATED = "status-updated",
	GAME_COMPLETED = "game-completed",
	CARDS_DEALT = "cards-dealt"
}

export const namesConfig: Config = {
	dictionaries: [ names ],
	separator: " ",
	length: 1
};